--- packages/ui-components/src/search-bar.ts
@@ -1,17 +1,0 @@
-import { InputGroup } from "@blueprintjs/core";
-import h from "@macrostrat/hyper";
-
-export function SearchBar({ onChange, placeholder = "Search..." }) {
-  return h(InputGroup, {
-    className: "search-bar",
-    size: "large",
-    fill: true,
-    round: false,
-    placeholder,
-    onChange: (e) => {
-      const value = e.target.value;
-      onChange(value);
-    },
-    leftIcon: "search",
-  });
-}
